rename nethandlers as netextensions
