+++ netextensions/gossip.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2024, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package nethandlers
+package netextensions
 
 import (
 	"context"
